requirements/base.txt
# ===============================================
# DEPENDENCIAS BASE - SISTEMA CATEQUESIS
# ===============================================

# Framework principal
Flask==2.3.3
Werkzeug==2.3.7

# Base de datos y ORM
pyodbc==4.0.39
pymssql==2.2.8
SQLAlchemy==2.0.23

# Validación y serialización
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Autenticación y seguridad
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
cryptography==41.0.7

# CORS y middleware
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Variables de entorno
python-dotenv==1.0.0

# Utilidades de fecha y tiempo
python-dateutil==2.8.2
pytz==2023.3

# HTTP y requests
requests==2.31.0
urllib3==2.0.7

# Logging mejorado
colorlog==6.7.0

# Validación de datos
email-validator==2.1.0
phonenumbers==8.13.25

# Generación de PDFs
reportlab==4.0.5
WeasyPrint==60.1

# Envío de emails
Flask-Mail==0.9.1

# Manejo de archivos
openpyxl==3.1.2
python-magic==0.4.27

# Cache
Flask-Caching==2.1.0

# Paginación
Flask-SQLAlchemy==3.0.5

# Utilidades generales
click==8.1.7
six==1.16.0
packaging==23.2